Fix milestone dialog background color
