+++ app/src/main/java/journal/gratitude/com/gratitudejournal/di/SettingsModule.kt
@@ -1,0 +1,8 @@
+package journal.gratitude.com.gratitudejournal.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+import journal.gratitude.com.gratitudejournal.repository.EntryRepository
+import journal.gratitude.com.gratitudejournal.repository.EntryRepositoryImpl
+import journal.gratitude.com.gratitudejournal.ui.settings.SettingsFragment
