+++ app/src/main/java/journal/gratitude/com/gratitudejournal/ui/timeline/TimelineFragment.kt
@@ -1,28 +1,22 @@
 package journal.gratitude.com.gratitudejournal.ui.timeline
 
-import android.Manifest
 import android.app.Activity.RESULT_OK
 import android.content.ActivityNotFoundException
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.net.Uri
 import android.os.Bundle
-import android.os.Environment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.PopupMenu
 import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
-import androidx.appcompat.app.AlertDialog
-import androidx.core.app.ActivityCompat
 import androidx.core.content.FileProvider
 import androidx.core.os.bundleOf
 import androidx.core.view.isVisible
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.FragmentNavigatorExtras
 import androidx.navigation.fragment.findNavController
 import com.crashlytics.android.Crashlytics
@@ -37,14 +31,13 @@ import journal.gratitude.com.gratitudejournal.ui.calendar.EntryCalendarListener
 import journal.gratitude.com.gratitudejournal.ui.entry.EntryFragment.Companion.DATE
 import journal.gratitude.com.gratitudejournal.ui.entry.EntryFragment.Companion.IS_NEW_ENTRY
 import journal.gratitude.com.gratitudejournal.ui.entry.EntryFragment.Companion.NUM_ENTRIES
-import journal.gratitude.com.gratitudejournal.util.backups.*
+import journal.gratitude.com.gratitudejournal.util.backups.CSVReaderImpl
+import journal.gratitude.com.gratitudejournal.util.backups.ExportCallback
+import journal.gratitude.com.gratitudejournal.util.backups.parseCsv
 import journal.gratitude.com.gratitudejournal.util.toLocalDate
 import kotlinx.android.synthetic.main.timeline_fragment.*
-import kotlinx.coroutines.launch
 import org.threeten.bp.LocalDate
-import org.threeten.bp.LocalDateTime
 import java.io.File
-import java.io.FileWriter
 import java.io.InputStream
 import java.util.*
 import javax.inject.Inject
@@ -132,14 +125,6 @@ class TimelineFragment : DaggerFragment() {
                             openContactForm()
                             true
                         }
-                        R.id.export_data -> {
-                            exportData()
-                            true
-                        }
-                        R.id.import_data -> {
-                            importData()
-                            true
-                        }
                         else -> false
                     }
                 }
@@ -211,41 +196,7 @@ class TimelineFragment : DaggerFragment() {
         }
     }
 
-    private fun importData() {
-        val alertDialog: AlertDialog? = activity?.let {
-            val builder = AlertDialog.Builder(it)
-            builder.apply {
-                setTitle(R.string.import_data_dialog)
-                setMessage(R.string.import_data_dialog_message)
-                setPositiveButton(R.string.ok) { dialog, id ->
-                    selectCSVFile()
-                }
-                setNegativeButton(R.string.cancel) { _, _ -> }
-            }
-            // Create the AlertDialog
-            builder.create()
-        }
-        alertDialog?.show()
-    }
 
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>, grantResults: IntArray
-    ) {
-        when (requestCode) {
-            MY_PERMISSIONS_REQUEST_WRITE_EXTERNAL -> {
-                // If request is cancelled, the result arrays are empty.
-                if ((grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
-                    performExport()
-                } else {
-                    Toast.makeText(context, R.string.permission_export, Toast.LENGTH_SHORT).show()
-                }
-                return
-            }
-            else -> {
-            }
-        }
-    }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         when (requestCode) {
@@ -284,62 +235,6 @@ class TimelineFragment : DaggerFragment() {
             Crashlytics.logException(exception)
 
             Toast.makeText(context, R.string.error_parsing, Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    private fun exportData() {
-        //TODO consolidate with other exportDB calls
-
-        val permission = ActivityCompat.checkSelfPermission(
-            activity!!,
-            Manifest.permission.WRITE_EXTERNAL_STORAGE
-        )
-
-        if (permission != PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(
-                arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE),
-                MY_PERMISSIONS_REQUEST_WRITE_EXTERNAL
-            )
-        } else {
-            performExport()
-        }
-    }
-
-    private fun performExport() {
-        firebaseAnalytics.logEvent(EXPORTED_DATA, null)
-
-        val dir =
-            Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-        if (!dir.exists()) {
-            dir.mkdir()
-        }
-        val date = LocalDateTime.now().withNano(0).toString().replace(':', '-')
-        val file = File(dir, "PresentlyBackup$date.csv")
-
-        val fileExporter = FileExporter(CSVWriterImpl(FileWriter(file)))
-
-        lifecycleScope.launch {
-            when (val result = fileExporter.exportToCSV(viewModel.getTimelineItems(), file)) {
-                is CsvCreated -> exportCallback.onSuccess(result.file)
-                is CsvError -> exportCallback.onFailure(result.exception)
-            }
-        }
-    }
-
-    private fun selectCSVFile() {
-        firebaseAnalytics.logEvent(LOOKED_FOR_DATA, null)
-
-        val intent = Intent(Intent.ACTION_GET_CONTENT)
-//        intent.addCategory(Intent.CATEGORY_OPENABLE)
-        intent.type = "text/csv|text/comma-separated-values|application/csv"
-        val mimeTypes = arrayOf("text/comma-separated-values", "text/csv")
-        intent.putExtra(Intent.EXTRA_MIME_TYPES, mimeTypes)
-
-        try {
-            startActivityForResult(Intent.createChooser(intent, "Select"), IMPORT_CSV)
-        } catch (ex: ActivityNotFoundException) {
-            Crashlytics.logException(ex)
-            Toast.makeText(context, R.string.no_app_found, Toast.LENGTH_SHORT).show()
         }
     }
