+++ app/src/main/res/menu/overflow_menu.xml
@@ -6,10 +6,4 @@
     <item
         android:id="@+id/contact_us"
         android:title="@string/contact_us" />
-    <item
-        android:id="@+id/export_data"
-        android:title="@string/back_up_entries_to_csv" />
-    <item
-        android:id="@+id/import_data"
-        android:title="@string/import_entries_from_backup" />
 </menu>
