+++ app/src/main/res/values/colors.xml
@@ -304,7 +304,7 @@
     <color name="katieAndroidWidgetColor">#4D1616</color>
     <color name="katieBackgroundColor">#FEE9E6</color>
     <color name="katieMainTextAndButtonColor">#241E1E</color>
-    <color name="katieHintQuoteTextColor">#4B4343</color>
+    <color name="katieHintQuoteTextColor">#706767</color>
     <color name="katieDateTextEntryScreenTextColor">#241E1E</color>
     <color name="katieTimelineColor">#3B1011</color>
     <color name="katieToolbarColor">#FEC2B8</color>
