+++ app/src/main/res/values/strings.xml
@@ -78,4 +78,8 @@
     <string name="dropbox_summary_on">Your entries are being backed up to Dropbox</string>
     <string name="dropbox_setting_title">Back up your data with Dropbox</string>
     <string name="dropbox_backup_frequency">Dropbox Backup Frequency</string>
+    <string name="one_time_export_summary">Manually export your entries to CSV format in your Downloads folder</string>
+    <string name="one_time_export">Export entries to CSV</string>
+    <string name="import_summary">Select a backed up CSV file to import</string>
+
 </resources>
